% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ercotr-data.r
\docType{data}
\name{ercot_palette}
\alias{ercot_palette}
\title{Palette data for this package}
\format{A character vector.}
\usage{
ercot_palette
}
\description{
Data used by the `scale_x_ercot()` functions.
}
\keyword{datasets}
