% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ercotr-data.r
\docType{data}
\name{ercot_palette2}
\alias{ercot_palette2}
\title{Palette data for this package}
\format{A character vector.}
\usage{
ercot_palette2
}
\description{
Data used by the `scale_x_ercot2()` functions.
}
\keyword{datasets}
