% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color}
\alias{scale_color}
\alias{get_ercot_palette_scaled}
\alias{get_ercot_palette}
\title{get_ercot_palette}
\usage{
scale_color(r, g, b, max = 255, scale = 1)

get_ercot_palette_scaled(num_colors = 12, max = 255, scale = 1,
  with_contrast = TRUE)

get_ercot_palette(num_colors = 12, with_contrast = TRUE)
}
\arguments{
\item{r}{red number, Max: 255}

\item{g}{green number Max: 255}

\item{b}{blue number Max: 255}

\item{max}{max number, Default: 255}

\item{scale}{number between 0 and 1, Default: 1}

\item{num_colors}{number between 0 and 12 Default: 12}

\item{with_contrast}{boolean, Default: TRUE
(TRUE provides maroon as secondary color, FALSE provides grey
as secondary color)}
}
\value{
A character vector with colors.
}
\description{
Used to change the "luminance" of colors for `get_ercot_palette_scaled()`,
(athough it really is a general function that can be used for other purposes as well).

Provides greater customability than `get_ercot_palette()`.

Used for `ercot_pal()`.
}
\details{
See `get_ercot_palette()` and `get_ercot_palette_scaled()`.

See `get_ercot_palette()`.
}
