% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_ercot.R
\name{ercot_pal}
\alias{ercot_pal}
\alias{scale_color_ercot}
\alias{scale_fill_ercot}
\title{ercot_pal}
\usage{
ercot_pal()

scale_color_ercot(...)

scale_fill_ercot(...)
}
\arguments{
\item{...}{dots}
}
\description{
with contrast
}
