% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_ercot.R
\name{ercot_pal2}
\alias{ercot_pal2}
\alias{scale_color_ercot2}
\alias{scale_fill_ercot2}
\title{ercot_pal2}
\usage{
ercot_pal2()

scale_color_ercot2(...)

scale_fill_ercot2(...)
}
\arguments{
\item{...}{dots}
}
\description{
withOUT contrast
}
